requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
dj-database-url==0.5.0
django-redis
django-filter==21.1
django-ipware==3.0.2
django-model-utils==4.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.7.0
